requirements.txt
pandas>=1.3.0
beautifulsoup4>=4.9.0
lxml>=4.6.0
requests>=2.25.0
python-dotenv>=0.19.0